+++ LuckyWheel/src/main/java/com/example/LuckyWheel/feature/wheel/dto/WheelDTO.java
@@ -17,7 +17,7 @@ public class WheelDTO {
     private boolean active;
 
     @JsonProperty("resourceRequire")
-    private Integer resourceRequire;
+    private Long resourceRequire;
 
     @JsonProperty("listGiftRandom")
     private List<GiftRandomDTO> listGiftRandom;
